Relocate QuickLauncher, update imports
